Make every /merge wait cancellable within ~100ms
Cancellation was only checked at the top of each poll iteration, while the
poll delay itself was an uninterruptible setTimeout. A cancel could not take
effect for up to a full poll interval: 30s for target-branch CI, and up to
5 minutes in waitForPRReady.

- Add cancellableSleep() to interruptible-sleep.lib.mjs, which polls an
  isCancelled predicate every 100ms and still honours SIGINT/SIGTERM.
- Route every merge wait loop through it.
- Make MergeQueueProcessor.sleep() the single cancellable primitive, so all
  call sites are cancellable and existing test stubs keep working. This
  supersedes the separate cancellableSleep helper from #1807.
- Thread isCancelled into checkPRMergeable so the UNKNOWN-mergeability retry
  delay aborts early, and skip (not fail) a PR cancelled mid-check.
- Guard the timeout paths that made an extra API round-trip after a cancel.
- Split waitForCI/waitForBranchCI into github-merge-ci-wait.lib.mjs to stay
  under the 1500-line limit; re-exported for existing importers.

Refs #2072
