+++ src/interruptible-sleep.lib.mjs
@@ -49,4 +49,37 @@ export function interruptibleSleep(ms) {
   });
 }
 
-export default { interruptibleSleep };
+/**
+ * Sleep for `ms` milliseconds, resolving early if SIGINT/SIGTERM arrives or if
+ * `isCancelled()` starts returning true.
+ *
+ * Issue #2072: polling loops used to `await new Promise(r => setTimeout(r, pollInterval))`
+ * and only re-check cancellation on the next iteration. With a 30s poll interval that
+ * made `/merge` keep running for up to a full interval after the Cancel button was
+ * pressed. Sleeping in short steps lets cancellation take effect within `stepMs`.
+ *
+ * @param {number} ms - Duration in milliseconds
+ * @param {Function|null} isCancelled - Predicate polled during the sleep
+ * @param {Object} [options]
+ * @param {number} [options.stepMs=100] - Granularity at which `isCancelled` is polled
+ * @returns {Promise<{interrupted: boolean, cancelled: boolean}>}
+ */
+export async function cancellableSleep(ms, isCancelled = null, options = {}) {
+  const { stepMs = 100 } = options;
+
+  if (!isCancelled) {
+    const { interrupted } = await interruptibleSleep(ms);
+    return { interrupted, cancelled: false };
+  }
+
+  const deadline = Date.now() + ms;
+  while (Date.now() < deadline) {
+    if (isCancelled()) return { interrupted: false, cancelled: true };
+    const { interrupted } = await interruptibleSleep(Math.min(stepMs, deadline - Date.now()));
+    if (interrupted) return { interrupted: true, cancelled: isCancelled() };
+  }
+
+  return { interrupted: false, cancelled: isCancelled() };
+}
+
+export default { interruptibleSleep, cancellableSleep };
